Add conversion-map and aggregate-converter types
- Adds interfaces and standard implementations for 2 types:

-- A dictionary-like generic collection that has Type for its key type,
and IConverter<T, U> for its value type. The T (convert-from type) is
the same for all values and is specified by the collection instance's
own type parameter. The U (convert-to type) is specified by a type
parameter on the accessor method.

-- An aggregate converter type that converts from values of T (type
parameter on the instance) to U (type parameter on the convert method).
An implementation is provided that uses the map described above to
provide conversion functions for the type specified by the call.

Both of the types implementations and interfaces have terrible names.
I'm committing them as is in hopes that someone else will learn the
behaviour and come up with a better name than I have.